extracted Fake tests to their own suite
